api/requirements.txt
Flask
Flask-PyMongo
flask-cors
python-dotenv
marshmallow
flask-jwt-extended
bcrypt
